+++ library/src/main/java/mobi/parchment/widget/adapterview/AnimationFrameScheduler.java
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (C) 2014 Emir Hasanbegovic and Parchment contributors.
+
+package mobi.parchment.widget.adapterview;
+
+public interface AnimationFrameScheduler {
+    void requestAnimationFrame();
+}